Token Hud Updates
* Create the TokenHUD5e class
* Incorperate core Foundry HUD html into the project
* Added Armor Class Icon to the HUD
* Incorperated condition changes into the new TokenHUD5e class
* Fixes toggle for Generic 5 condition
