+++ public/templates/hud/token-hud.hbs
@@ -1,0 +1,132 @@
+<form
+    id="{{id}}"
+    class="{{classes}}"
+    data-appid="{{appId}}"
+    autocomplete="off"
+    onsubmit="event.preventDefault();"
+>
+
+    <div class="col left">
+        <div class="attribute elevation">
+            <i class="fas fa-angle-up"></i>
+            <input
+                type="text"
+                name="elevation"
+                value="{{elevation}}"
+            >
+        </div>
+
+        <div class="a5e-armorclass">
+            {{AC}}
+        </div>
+
+        {{#if canConfigure}}
+            <div
+                class="control-icon"
+                data-action="config"
+            >
+                <i class="fas fa-cog"></i>
+            </div>
+        {{/if}}
+
+        <div
+            class="control-icon {{targetClass}}"
+            data-action="target"
+        >
+            <i class="fas fa-bullseye"></i>
+        </div>
+    </div>
+
+    <div class="col middle">
+        <div class="attribute bar2">
+            {{#if displayBar2}}
+                <input
+                    type="text"
+                    name="bar2.value"
+                    data-bar="bar2"
+                    value="{{bar2Data.value}}"
+                    {{#unless
+                    bar2Data.editable}}
+                    disabled
+                    {{/unless}}
+                >
+            {{/if}}
+        </div>
+
+        <div class="attribute bar1">
+            {{#if displayBar1}}
+                <input
+                    type="text"
+                    name="bar1.value"
+                    data-bar="bar1"
+                    value="{{bar1Data.value}}"
+                    {{#unless
+                    bar1Data.editable}}
+                    disabled
+                    {{/unless}}
+                >
+            {{/if}}
+        </div>
+    </div>
+
+    <div class="col right">
+        {{#if isGM}}
+            <div
+                class="control-icon {{visibilityClass}}"
+                data-action="visibility"
+            >
+                <img
+                    src="{{icons.visibility}}"
+                    width="36"
+                    height="36"
+                    title="{{localize 'HUD.ToggleVis'}}"
+                />
+            </div>
+        {{/if}}
+
+        <div
+            class="control-icon {{effectsClass}}"
+            data-action="effects"
+        >
+            <img
+                src="{{icons.effects}}"
+                width="36"
+                height="36"
+                title="{{localize 'HUD.AssignStatusEffects'}}"
+            />
+            <div class="status-effects">
+                {{#each statusEffects as |status src|}}
+                    <div class="condition-container">
+                        <img
+                            class="effect-control {{status.cssClass}}"
+                            src="{{status.src}}"
+                            {{#if
+                            status.title}}
+                            title="{{status.title}}"
+                            {{/if}}
+                            data-status-id="{{status.id}}"
+                        />
+                        <div class="condition-name">{{status.title}}</div>
+                    </div>
+                {{/each}}
+                <div class="clear-all-conditions">
+                    <i class="fa-solid fa-octagon-xmark"></i> {{localize 'A5E.UIClearAll'}}
+                </div>
+            </div>
+        </div>
+
+        {{#if canToggleCombat}}
+            <div
+                class="control-icon {{combatClass}}"
+                data-action="combat"
+            >
+                <img
+                    src="{{icons.combat}}"
+                    width="36"
+                    height="36"
+                    title="{{localize 'HUD.ToggleCombatState'}}"
+                />
+            </div>
+        {{/if}}
+    </div>
+</form>
