+++ src/scss/components/_hud-armorclass.scss
@@ -1,0 +1,14 @@
+.placeable-hud .a5e-armorclass {
+  border-radius: 50% 50% 50% 50% / 12% 12% 88% 88%;
+  box-sizing: content-box;
+  width: 45px;
+  line-height: 45px;
+  flex: 0 0 40px;
+  margin: 8px 0;
+  font-size: var(--font-size-28);
+  text-align: center;
+  color: var(--color-text-light-3);
+  background: rgba(0, 0, 0, 0.6);
+  box-shadow: 0 0 15px var(--color-shadow-dark);
+  border: 1px solid #333;
+}
