+++ src/scss/main.scss
@@ -20,6 +20,7 @@
 @import './components/header-shield';
 @import './components/health-bar';
 @import './components/hit-die';
+@import './components/hud-armorclass';
 @import './components/image';
 @import './components/input';
 @import './components/item';
